requirements.txt
python-dotenv==1.1.1
pydantic==2.11.9
pydantic-settings==2.10.1
crewai==0.186.1
fastapi==0.116.1
uvicorn==0.35.0
streamlit==1.49.1
chromadb==1.0.21
llama-index==0.14.0
llama-index-embeddings-huggingface==0.6.1
llama-index-llms-groq==0.4.1
llama-index-vector-stores-chroma==0.5.3
sacrebleu
rouge-score
sentence-transformers